+++ sandbox/modules/identity/src/internalClusterTest/java/org/opensearch/identity/UserPersistenceIT.java
@@ -9,11 +9,6 @@
 package org.opensearch.identity;
 
 import org.junit.rules.TemporaryFolder;
-import org.opensearch.action.admin.cluster.health.ClusterHealthResponse;
-import org.opensearch.action.admin.cluster.node.info.NodeInfo;
-import org.opensearch.action.admin.cluster.state.ClusterStateResponse;
-import org.opensearch.cluster.health.ClusterHealthStatus;
-import org.opensearch.cluster.health.ClusterIndexHealth;
 import org.opensearch.common.SuppressForbidden;
 import org.opensearch.plugins.Plugin;
 import org.opensearch.test.OpenSearchIntegTestCase;
@@ -26,9 +21,6 @@ import java.util.Collection;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import static org.hamcrest.Matchers.equalTo;
-import static org.opensearch.test.hamcrest.OpenSearchAssertions.assertNoTimeout;
-
 @OpenSearchIntegTestCase.ClusterScope(scope = OpenSearchIntegTestCase.Scope.TEST, numDataNodes = 0)
 public class UserPersistenceIT extends HttpSmokeTestCaseWithIdentity {
 
@@ -36,37 +28,6 @@ public class UserPersistenceIT extends HttpSmokeTestCaseWithIdentity {
     protected Collection<Class<? extends Plugin>> nodePlugins() {
         List<Class<? extends Plugin>> plugins = super.nodePlugins().stream().collect(Collectors.toList());
         return plugins;
-    }
-
-    public void startNodes() throws Exception {
-        final String clusterManagerNode = internalCluster().startClusterManagerOnlyNode();
-
-        ClusterStateResponse clusterStateResponse = client(clusterManagerNode).admin()
-            .cluster()
-            .prepareState()
-            .setClusterManagerNodeTimeout("1s")
-            .clear()
-            .setNodes(true)
-            .get();
-        assertNotNull(clusterStateResponse.getState().nodes().getClusterManagerNodeId());
-
-        // start another node
-        final String dataNode = internalCluster().startDataOnlyNode();
-        clusterStateResponse = client(dataNode).admin()
-            .cluster()
-            .prepareState()
-            .setClusterManagerNodeTimeout("1s")
-            .clear()
-            .setNodes(true)
-            .setLocal(true)
-            .get();
-        assertNotNull(clusterStateResponse.getState().nodes().getClusterManagerNodeId());
-        // wait for the cluster to form
-        assertNoTimeout(client().admin().cluster().prepareHealth().setWaitForNodes(Integer.toString(2)).get());
-        List<NodeInfo> nodeInfos = client().admin().cluster().prepareNodesInfo().get().getNodes();
-        assertEquals(2, nodeInfos.size());
-
-        Thread.sleep(3000);
     }
 
     @SuppressForbidden(reason = "manipulates system properties for testing")
@@ -115,19 +76,7 @@ public class UserPersistenceIT extends HttpSmokeTestCaseWithIdentity {
 
             startNodes();
 
-            ClusterHealthResponse clusterHealthResponse = client().admin()
-                .cluster()
-                .prepareHealth()
-                .setClusterManagerNodeTimeout("1s")
-                .get();
-
-            assertTrue(
-                ConfigConstants.IDENTITY_DEFAULT_CONFIG_INDEX + " index exists",
-                clusterHealthResponse.getIndices().containsKey(ConfigConstants.IDENTITY_DEFAULT_CONFIG_INDEX)
-            );
-
-            ClusterIndexHealth identityIndexHealth = clusterHealthResponse.getIndices().get(ConfigConstants.IDENTITY_DEFAULT_CONFIG_INDEX);
-            assertEquals(ClusterHealthStatus.GREEN, identityIndexHealth.getStatus());
+            ensureIdentityIndexIsGreen();
         } catch (IOException ioe) {
             fail("error creating temporary test file in " + this.getClass().getSimpleName());
         }
@@ -165,19 +114,7 @@ public class UserPersistenceIT extends HttpSmokeTestCaseWithIdentity {
 
             startNodes();
 
-            ClusterHealthResponse clusterHealthResponse = client().admin()
-                .cluster()
-                .prepareHealth()
-                .setClusterManagerNodeTimeout("1s")
-                .get();
-
-            assertTrue(
-                ConfigConstants.IDENTITY_DEFAULT_CONFIG_INDEX + " index exists",
-                clusterHealthResponse.getIndices().containsKey(ConfigConstants.IDENTITY_DEFAULT_CONFIG_INDEX)
-            );
-
-            ClusterIndexHealth identityIndexHealth = clusterHealthResponse.getIndices().get(ConfigConstants.IDENTITY_DEFAULT_CONFIG_INDEX);
-            assertThat(identityIndexHealth.getStatus(), equalTo(ClusterHealthStatus.GREEN));
+            ensureIdentityIndexIsGreen();
         } catch (IOException ioe) {
             fail("error creating temporary test file in " + this.getClass().getSimpleName());
         }